+++ src/spring/HISTORY.md
@@ -1,5 +1,14 @@
 Release History
 ===============
+1.8.1
+---
+* Support `--ca-cert-name` and `--sync-cert` in `az spring application-accelerator customized-accelerator`.
+* Support client cert validation for customized accelerator with CA certificate.
+
+1.8.0
+---
+* Add Azure Spring Apps StandardGen2 tier.
+
 1.7.3
 ---
 * Fix `subscription_id` AAZSimpleValue type error